==> 3 - Random Forest.ipynb <==
## 3 - Random Forest

import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn import cross_validation
from sklearn.grid_search import GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.feature_selection import RFECV
from operator import itemgetter

train_data = pd.read_csv('./input/train_prep.csv')
test_data = pd.read_csv('./input/test_prep.csv')

param = train_data.ix[:, :-1] # take all rows and all but last column from training data
res = train_data.ix[:, -1] # take all rows and only last column from training data
test_data = test_data.drop('ID', 1) # drop ID column from test data

# create output file
def create_output_file(data, file_name):
    output = []
    for pred in data:
        oi = [0] * 5
        oi[pred - 1] = 1
        output.append(oi)
    output = pd.DataFrame(output, columns=['Adoption', 'Died', 'Euthanasia', 'Return_to_owner', 'Transfer'])
    output.index.names = ['ID']
    output.index += 1
    output.to_csv('./output/Sub-' + file_name, index_label='ID')

rfc = RandomForestClassifier(n_estimators = 200, max_features='auto') # auto = sqrt(max_features)
# rfc = RandomForestClassifier(n_estimators = 100, max_features=None) # None = max_features
cross_validation.cross_val_score(rfc, param, res)

rfc.fit(param, res) # fit the data
predictions = rfc.predict(test_data) # make predictions
create_output_file(predictions, 'RF.csv')
print predictions

### With Pipeline and GridSerchCV

pipeline = Pipeline([
            ('featureSelection', RFECV(estimator=RandomForestClassifier(n_estimators=20), scoring='neg_log_loss')),
            ('clf', RandomForestClassifier(n_estimators=100))
        ])
params = {
    "clf__max_depth": [2, 3, None],
    "clf__max_features": [0.75, 1.0],
    "clf__min_samples_split": [2, 3, 10],
    "clf__min_samples_leaf": [2, 3, 10],
    "clf__bootstrap": [True, False],
    "clf__criterion": ["gini", "entropy"]
}

def report(grid_scores, n_top=3):
    top_scores = sorted(grid_scores, key=itemgetter(1), reverse=True)[:n_top]
    for i, score in enumerate(top_scores):
        print("Model with rank: {0}".format(i + 1))
        print("Mean validation score: {0:.3f} (std: {1:.3f})".format(
              score.mean_validation_score,
              np.std(score.cv_validation_scores)))
        print("Parameters: {0}".format(score.parameters))
        print("")

grid_search = GridSearchCV(pipeline, params, n_jobs=-1)
grid_search.fit(param, res)
report(grid_search.grid_scores_)

predictions = grid_search.predict_proba(test_data)
grid_search.predict_proba(test_data)[:5]

predictions = grid_search.predict(test_data)
create_output_file(predictions, 'RF-gridSearch.csv')
