==> 6 - SVM.ipynb <==
## SVM 

from sklearn.svm import SVC
from sklearn import cross_validation
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.pipeline import make_pipeline
import pandas as pd

train_data = pd.read_csv('./input/train_prep.csv')
test_data = pd.read_csv('./input/test_prep.csv')

param = train_data.ix[:, :-1] # take all rows and all but last column from training data
res = train_data.ix[:, -1] # take all rows and only last column from training data
test_data = test_data.drop('ID', 1) # drop ID column from test data

# create output file
def create_output_file(data, file_name):
    output = []
    for pred in data:
        oi = [0] * 5
        oi[pred - 1] = 1
        output.append(oi)
    output = pd.DataFrame(output, columns=['Adoption', 'Died', 'Euthanasia', 'Return_to_owner', 'Transfer'])
    output.index.names = ['ID']
    output.index += 1
    output.to_csv('./output/Sub-' + file_name, index_label='ID')

clf = SVC(probability=True)
cross_validation.cross_val_score(clf, param, res, scoring="neg_log_loss")

clf.fit(param, res)

pred = clf.predict(test_data)
create_output_file(pred, 'SVM.csv')
print pred

### Pipeline with feature selection and estimation

param_new = SelectKBest(chi2, k=3).fit_transform(param, res)
cross_validation.cross_val_score(clf, param_new, res)

selector = SelectKBest(chi2, k=3)
predictor = make_pipeline(selector, clf)
predictor.fit(param, res)
predictions = predictor.predict_proba(test_data)

predictions = predictor.predict(test_data)
create_output_file(predictions, 'SVM-Kbest.csv')
